Add STA-only Wi-Fi mode, extend password support, and fix ESP-IDF 5.5 build compatibility
Introduce Wi-Fi STA-only mode (WIFI_MODE=2) across Web UI, REST config/status handling, settings validation, and runtime mode switching while preserving existing AP/APSTA behavior and recovery flow.
Increase Wi-Fi password capacity to 64 characters for modern WPA passphrases, keep legacy settings migration binary-compatible by freezing old-struct password length, and improve API feedback by returning a correct password-related error message.
Fix ESP-IDF 5.5 toolchain compatibility issues by correcting FatFS f_getfree() argument type in sd_card code and replacing direct no-arg console command bindings with proper esp_console callback wrappers.
Bump firmware version string to 1.2.2.

+++ components/sd_card/esp_sd_card.c
@@ -303,7 +303,7 @@ uint32_t esp_sd_card_get_free_space()
             }
                 
             FATFS *fs;
-            uint64_t fre_clust; //, tot_sect;
+            DWORD fre_clust; //, tot_sect;
             DWORD fre_sect;
             /* Get volume information and free clusters of drive 0 */
             int res = f_getfree("/sdcard/", &fre_clust, &fs);
